feat(sankey): 品質スコアの詳細ダイアログを /sankey サイドパネルから開けるように
/quality の ScoreDetailDialog を共通コンポーネント化し、/sankey-svg の
品質スコアブロックのバッジ・「詳細」ボタンから同一ダイアログを開けるようにした。

- ScoreDetailDialog / 共通フォーマッタ(scoreColor,formatAmount,pct)を
  client/components/quality/ に切り出し、/quality と /sankey で共用（重複定義なし）
- /api/quality-scores/[pid] に ?full=1 を追加し、ダイアログに必要な全項目を返す
  （既定は従来どおり軽量プロジェクション）
- /sankey ではダイアログを createPortal で document.body に出し、背面サンキー図の
  wheel/pan ハンドラにイベントが伝播しないよう分離（ダイアログ内スクロール時に背面が動く不具合を修正）
- ダイアログをレスポンシブ対応：スコア行の折り返し、支出先テーブルの横スクロール収容、
  メトリクスのモバイル独立行化、セクション余白の縮小（スマホ表示の破綻を修正）
